New semantic analyzer: fix access to imported name in class body
We shouldn't use the line number of an imported name to decide
whether it's defined when the line number refers to another
file.
